Disable Ice modifier
There have been complaints about Ice modifier not being a good modifier as the upside wasn't noticeable, so people were deeming it a straight downgrade.
...well, that's because it apparently is.

The modifier's upside wouldn't work because it was modifying the victim's speed, as intended, but would try to recalculate the boss' speed instead. This has been "fixed".

That's not enough, as the function that calculates max speed uses a pre hook, rather than post. It'll always return `0.0` and make a player move at 520 hu/s if their speed multiplier is modified. 42 and I looked into this for a bit, but no matter what was done it'd crash the server, so I suggest that the modifier should be disabled for now.