Changes to Sniper
Kinda gets rid of issue #113

PR #106 allows damage calculation for internal stuff be more accurate, which is helpful for bosses/modifiers/etc , but the specific look-for-minicrit-but-not-crit-damage case got hit in the crossfire, and since we wanted to simplify sniper primaries, this (un)fortunately is a good opportunity to do so, changes:

- Bonus damage for bodyshots and tracer rounds on all sniper rifles have been removed
    - Instead, they're minicrit buffed. Includes huntsman.

- Bazaar bargain has its outline downside removed, as it simply didn't deserve it

It's worth mentioning that headshots, the main gimmick of the Sniper class of the Team Fortress 2 videogame, exist! Making use of them will deservingly increase your damage output.