Baby Face's Blaster: increase damage req to 400
In VSH, this weapon's charge meter is much easier to build as the threats for its downside simply do not exist, making the default speed gain feel annoying and undeserved.

Instead of trying to reinvent the wheel, the speed boost will remain a feature of the weapon, it'll just be a bit harder to get. :)