Allow Vagineer to have more than one sentry up on rage
Right now, if Vagineer rages while already having an active sentry, it'll be destroyed so another one can be built. That's pretty harsh for the boss as his rage doesn't have much going on defensively besides a little-ranged stun and with this happening commonly in a panic-raging scenario, he's usually getting shredded.

This PR makes it so the previous sentry won't be destroyed while allowing ownership of multiple sentries.