Quick Fix: decrease über rate and AoE radius
breaking the month-long no change streak!

the Quick-Fix has been getting a lot of use compared to the other mediguns, so to give them a better chance, I'm suggesting to remove the extra über rate bonus and lower the radius of the AoE buff to match the Kritzkrieg's. (the Kritzkrieg is in a pretty fine spot, imo.)