Degreaser: undo current attributes
This is a major change.

being completely honest with you, I don't think the Degreaser we have right now is healthy for the gamemode. being sent across a room (or half the map if the pyro is standing on a high edge), per pyro, is not a fun thing for bosses to deal with.

that said, the Degreaser with no changes is pretty shit in this gamemode so this is some sort of discussion point: how to make the Degreaser better?

if anyone has a suggestion, let us know somewhere, but this PR has a suggestion of my own: give it the passive weapon switch speed that it had until Tough Break. it's pretty tame and doesn't sound like much, but people could hopefully have an easier time comboing it with something, as it's the point of the Degreaser.