Tomislav: flatten custom stats
It doesn't make much sense to have custom positives and negatives that don't complement each other. In this case, the negative even makes the weapon pretty unusable.

This PR reduces stat spam and changes them into a neat, single one that enhances what the weapon is supposed to do. I think it'll still not be quite as good as the brass beast, but this at least makes the weapon possible to use.